requirements.txt
streamlit
pandas
folium
streamlit-folium
fpdf2
streamlit-lottie